+++ js/model3D/static/Model3D.svelte
@@ -2,13 +2,14 @@
 	import type { FileData } from "@gradio/upload";
 	import { BlockLabel, IconButton } from "@gradio/atoms";
 	import { File, Download } from "@gradio/icons";
+	import {add_new_model} from "../shared/utils"
 	import { _ } from "svelte-i18n";
 	import { onMount } from "svelte";
 	import * as BABYLON from "babylonjs";
 	import * as BABYLON_LOADERS from "babylonjs-loaders";
 
 	export let value: FileData | null;
-	export let clearColor: [number, number, number, number] = [0, 0, 0, 0];
+	export let clear_color: [number, number, number, number] = [0, 0, 0, 0];
 	export let label = "";
 	export let show_label: boolean;
 
@@ -24,7 +25,6 @@
 	let canvas: HTMLCanvasElement;
 	let scene: BABYLON.Scene;
 	let engine: BABYLON.Engine | null;
-
 	let mounted = false;
 
 	onMount(() => {
@@ -53,58 +53,7 @@
 				engine?.resize();
 			});
 		}
-		addNewModel();
-	}
-
-	function addNewModel(): void {
-		scene = new BABYLON.Scene(engine!);
-		scene.createDefaultCameraOrLight();
-
-		scene.clearColor = new BABYLON.Color4(...clearColor);
-
-		engine?.runRenderLoop(() => {
-			scene.render();
-		});
-
-		if (!value) return;
-
-		let url: string;
-		if (value.is_file) {
-			url = value.data;
-		} else {
-			let base64_model_content = value.data;
-			let raw_content = BABYLON.Tools.DecodeBase64(base64_model_content);
-			let blob = new Blob([raw_content]);
-			url = URL.createObjectURL(blob);
-		}
-
-		BABYLON.SceneLoader.ShowLoadingScreen = false;
-
-		BABYLON.SceneLoader.Append(
-			"",
-			url,
-			scene,
-			() => {
-				// scene.createDefaultCamera(createArcRotateCamera, replace, attachCameraControls)
-				scene.createDefaultCamera(true, true, true);
-				// scene.activeCamera has to be an ArcRotateCamera if the call succeeds,
-				// we assume it does
-				var helperCamera = scene.activeCamera! as BABYLON.ArcRotateCamera;
-
-				if (camera_position[0] !== null) {
-					helperCamera.alpha = (Math.PI * camera_position[0]) / 180;
-				}
-				if (camera_position[1] !== null) {
-					helperCamera.beta = (Math.PI * camera_position[1]) / 180;
-				}
-				if (camera_position[2] !== null) {
-					helperCamera.radius = camera_position[2];
-				}
-			},
-			undefined,
-			undefined,
-			"." + value["name"].split(".")[1]
-		);
+		add_new_model(canvas, scene, engine, value, clear_color, camera_position);
 	}
 </script>
